Correct pinouts on README.md
For the SmartShunt, the TX pin seems to be next to the VCC pin rather than next to GND as was shown in the first drawing (the second drawing for the UART-TTL was correct). Additionally, the orientation of the UART-TTL jack drawing has been rotated to reflect the orientation on the Victron SmartShunt. The required UART section was also added to the sample yaml configs.